Create fulltrainprune.py
This is the experiment we intend to run and check if we achieve a better accuracy and better efficiency by training it firstly with all the weights at initialization that we accept as a tradeoff. And then reinitialize weights after pruning as the weights we got after full training. So the entire process is:

1) Train the full network with all the data and obtain the final weights θ
2) Use train-prune loops to obtain the mask.
3) Once the mask m is obtained: use the θ*m as the weights for the network.